housekeeping: Implement automatic purging of trash://
This commit reimplements the 'empty trash' functionality not
to rely on nautilus, and adds an 'age' parameter.

We install timeouts that check once every hour and purges
trash and temporary files if they are older than the 'purge-after'
setting allows. The timeouts only act if the 'purge-trash' or
'purge-temp-files' settings are TRUE.

Temporary files are looked for in g_get_tmp_dir(), /tmp and /var/tmp.

https://bugzilla.gnome.org/show_bug.cgi?id=149572