power: Don't dim when idle is inhibited
We expect movie players and similar to inhibit idle as a way to
prevent the screensaver from kicking in. We may even get firefox
to do this for HTML5 video. But constant screen dimming is just as
disruptive to the movie experience, so I think we should not do it
when idle is inhibited.

https://bugzilla.gnome.org/show_bug.cgi?id=691002