power: Reset IDLETIME on resume
We used to do this using the SimulateUserActivity D-Bus function
from gnome-screensaver, but that doesn't exist anymore with
gnome-shell and the code was removed in commit
87a2490dc8c154a3eb7ead55c6a60e9e7fb81d1a

We shouldn't need to do it, but Xorg won't do it
for us (yet? https://bugs.freedesktop.org/show_bug.cgi?id=59576)

Reset the IDLETIME by poking left-shift programmatically on resume.