power: Fix blanking not working
When we got the ActiveChanged() signal from gnome-shell's screen lock,
we would reconfigure the blank time out to be much shorter. The problem
is that the timeout is based on idleness, and we would already have been
past that amount of idleness when we setup the new idle monitor.

Instead of basing ourselves on idle timeouts, just wait for the
screensaver to come on and blank it. We'll handle idleness as normal
(so it's possible to cancel the login, and have the screen be blanked
quickly).

https://bugzilla.gnome.org/show_bug.cgi?id=691965