tests: Run under X.org with dummy driver
Run a private X.org instance with the dummy driver for the tests. The dummy
driver supports basic XRandR, but no DPMS, which is good enough at least for
testing most of "power", "keyboard", "multimedia", and others. This is
necessary until Xvfb supports XRandR, which was proposed in
<http://lists.x.org/archives/xorg-devel/2013-January/035114.html>.

This makes the tests insensitive against input events from the real X.org which
happen during the tests, and also allows us to run the tests in headless CI
server environments.

Some distros like Fedora install Xorg as suid root, so copy the binary into our
temporary workdir to drop the suid bit and run it as user.