media-keys: Refer key grabbing to the shell
As events are not delivered to the root window when the shell itself
has grabbed the keyboard, shortcuts currently won't work in "special"
modes like the overview. While the shell does offer a mechanism to
selectively allow keybindings in these modes, it is only available
for shortcuts grabbed by the shell itself. In order to allow other
processes to use this feature as well, the shell now exposes a
simple grab API on DBus; start using this API to fix shortcuts
during compositor grabs.

https://bugzilla.gnome.org/show_bug.cgi?id=693016