power: Handle the shell coming back with a different screensaver state
When the shell crashes while using the screensaver (or it needs to be
killed because it won't remove a black screen), we checked for the
state of the screensaver, but didn't do anything with it.

Now we use the same codepath and make sure we disable the screensaver's
aggressive timeout if gnome-shell isn't in screensaver mode anymore.

https://bugzilla.gnome.org/show_bug.cgi?id=696290