Avoid calling setenv after starting threads
In particular, we can't get the locale value out of GSettings
in order to set LC_PAPER etc, since calling into GSettings
initializes the dconf backend which in turn uses gdbus, which
starts a worker thread.

As a simple workaround, set up the locale environment in
a small wrapper script that then exec's the g-s-d binary.
https://bugzilla.gnome.org/show_bug.cgi?id=701322