power: Fix non-transient notifications sticking around
When replacing an existing notification with a more up-to-date one,
we were closing the existing notification, and creating a new one in its
place. As, to clean up when a notification is dismissed by hand, we hook
up to the "closed" signal, we ended up zero'ing the pointer to the just
shown notification and making it impossible to remove later.

eg.
- 1st notification created
- updated state comes in
- notify_close() called
- notification pointer is replaced with newer notification
- main loop returns
- "closed" signal is received for the 1st notification (ah!)
- original notification is unref'ed, and its location is zero'd
  by weak pointer call.

That location is that of the new notification as well.

https://bugzilla.gnome.org/show_bug.cgi?id=702047