mouse: allow force-disabling touchpad
The mouse plugin currently makes sure that at least one pointing device
is present. When the last external device is unplugged, it enables the
touchpad (if present). However, that logic also enables the touchpad
when it itself is "unplugged", for example by a hardware switch.

Fix this by not enabling the touchpad if that was the device that was
just removed.

Fixes LP: #1454950