Cherry-picked relevant bits from upstream From d432d4e7ab0e18102f1ae78603e933547f63e6d4 Mon Sep 17 00:00:00 2001 From: Carlos Garnacho <carlosg@gnome.org> Date: Wed, 17 Dec 2014 15:33:45 +0100 Subject: [PATCH] mouse: Use gsettings-desktop-schemas mouse/touchpad settings
Most settings have been moved there, so make this plugin use the new
schemas for the handled settings, and remain using the settings-daemon
schema for settings that can just be honored on X11.

In order to simplify the situation for the future, the management of
"middle-button-enabled", "horiz-scroll-enabled" and "disable-while-typing"
settings have been removed, and the sane default is applied on non-libinput
devices on X11.

https://bugzilla.gnome.org/show_bug.cgi?id=742593